Move Qwen full-attention trace out of main
